requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==0.28.0
werkzeug==3.0.1
flask-cors==4.0.0